refactor(email): share the one locale interpolation helper
Five byte-identical copies of the same four-line function, three of them
differing only by a suffix — replaceVarEV, replaceVarPR — invented to dodge a
name collision inside a single package. auth_handler.go and
email_verification.go carried two of them side by side.

It moves to pkg/email, which every caller already imports for the Email type,
and takes its one test with it. That test was the only coverage any of the
five copies had.
